+++ lib/screens/stats/stats.dart
@@ -1,11 +1,7 @@
-import 'package:budgetiser/screens/stats/lineChart.dart';
-import 'package:budgetiser/screens/stats/simple_text_stat.dart';
+import 'package:budgetiser/screens/stats/line_chart.dart';
 import 'package:budgetiser/shared/dataClasses/account.dart';
-import 'package:budgetiser/shared/dataClasses/transaction_category.dart';
 import 'package:budgetiser/shared/picker/account_picker.dart';
 import 'package:budgetiser/shared/picker/monthPicker.dart';
-import 'package:budgetiser/shared/picker/select_account.dart';
-import 'package:budgetiser/shared/picker/select_category.dart';
 import 'package:flutter/material.dart';
 import 'package:budgetiser/drawer.dart';
 
@@ -39,34 +35,37 @@ class _StatsState extends State<Stats> {
       body: Padding(
         padding: const EdgeInsets.all(8.0),
         child: Column(
-        children: [
-          MonthPicker(onDateChangedCallback: (DateTime time) => {
-            setState(() {
-              startDate = time;
-            })
-          }),
-          Row(
-            children: [
-              Expanded(
-                child: AccountPicker(
-                  onAccountPickedCallback: (List<Account> accounts) {
-                    setState(() {
-                      _selectedAccounts = accounts;
-                    });
-                  },
+          children: [
+            MonthPicker(
+              onDateChangedCallback: (DateTime time) => {
+                setState(() {
+                  startDate = time;
+                })
+              },
+            ),
+            Row(
+              children: [
+                Expanded(
+                  child: AccountPicker(
+                    onAccountPickedCallback: (List<Account> accounts) {
+                      setState(() {
+                        _selectedAccounts = accounts;
+                      });
+                    },
+                  ),
                 ),
+              ],
+            ),
+            const SizedBox(height: 16),
+            if (_selectedAccounts != null)
+              LineChartTest(
+                accounts: _selectedAccounts!,
+                startDate: DateTime(startDate.year, startDate.month, 1),
+                endDate: DateTime(startDate.year, startDate.month + 1, 0),
               ),
-            ],
-          ),
-          const SizedBox(height: 16),
-          if (_selectedAccounts != null)
-          LineChartTest(
-            accounts: _selectedAccounts!,
-            startDate: DateTime(startDate.year, startDate.month, 1),
-            endDate: DateTime(startDate.year, startDate.month+1, 0),
-          ),
-        ],
-      )
-    ));
+          ],
+        ),
+      ),
+    );
   }
 }
